Added example fuzzy logic files
Included two python files from SciKit Fuzzy's examples. The file
plotting the control system is a surface plot and not currently helpful.
The tiping problem using the new skfuzzy api was helpful and used to
recreate HW3 from soft computing.

HW3.py is a recreation of the matlab scripts used during softcomputing
with Dr. Cohen. The plot is stored for reference.