Updated gap_error, rules, and started run function
Updated the gap error membership functions following the discussion with
Anoop. The units are now in seconds. The rules have been updated per my
initial design. I started to edit the run function to take in the pandas
dataframe but stopped at line 164. Trying to split the dataframe per
vehicle before use. Line 130 should do the trick when used in a for
loop.