Working fuzzy test
Managed to get the fuzzy system to estimate the acceleration. The excel
sheet is parsed, the vehicle x position data is used to calculate the
gap, the gap error is calculated, and the rate the gap error changes is
calculated. The gap error and rate that changes is used as inputs to the
fuzzy system.