Test fuzzy controller iterate over output data (#27)
* Updated membership function and fuzzy logic

* Removed unecessay variable

* Working fuzzy test

Managed to get the fuzzy system to estimate the acceleration. The excel
sheet is parsed, the vehicle x position data is used to calculate the
gap, the gap error is calculated, and the rate the gap error changes is
calculated. The gap error and rate that changes is used as inputs to the
fuzzy system.

* saving vehicle FL in/out information to csv

* Reduced human driver top speed

* Updated ideal_gap to 1

* updated gap error rate calculation

* Updated Krauss model car max speed

* Updated gap calculation including vehicle length

* Set tau to 1

* Updated csv file in use to 004 fcdout

* Updated membership functions

Bringing the acceleration values closer to the CACC model when cruising
on the highway.

* Increased the SUMO lane length to 8 miles

The vehicles will stabilize over the first two miles. The assesment will
be conducted over the middle 4 miles. The final two miles will be left
for the vehicles to close out.

* Updated function structure

* Added cacc route and network files

* Created initial attempt to us FL live

* FL live is almost working

The FL is ran but the setSpeed does not appear to go through.

* Fixed the vehicle_gap and gap_error calculation

The vehicle gap had a unit of meters and it needed to be in seconds. The calculation for that has been updated. The gap error has been updated as well. Previously, the gap error rate was 0 no matter what. That has been updated too.

* Added gap error plot for Krauss model

Includes work adding the argument to run the krauss model instead of the
FIS.

* January 2023 status report

* Copied over and renamed working FIS

* Renamed sumo_script to test_sumo_script

* Cleaning up unstaged changes