Added the gap error difference from min gap error
It did not work well in the FIS. I need to sit down thinking about it
some more. Try adding it by itself in the rules but only if it ands
together with everything else in that rule? Not sure honestly. Also,
consider moving the (self.gap_error['Large'] &
self.gap_error_rate['Medium']) to the large acceleration rule.