Mem function addition (#32)
* Created basic system block diagram

* Added the gap error difference from min gap error

It did not work well in the FIS. I need to sit down thinking about it
some more. Try adding it by itself in the rules but only if it ands
together with everything else in that rule? Not sure honestly. Also,
consider moving the (self.gap_error['Large'] &
self.gap_error_rate['Medium']) to the large acceleration rule.

* Plotting gap error rate

* Removed emissions from recording

* Moved membership pair to large acceleration

Moved the large gap error & medium gap error rate pair from the medium
accel to the large accel. This did not change the result much.

* Save gap error and rate to csv