Fixed memory leak issue
The skfuzzy control simulation was created four times per SUMO time
step. Each time the vehicles were calculating the acceleration. The
skfuzzy control space creation was moved and occurs once per chromosome.

The skfuzzy control holds cache for up to 100 simulations. This can be
expanded now that the real memory issue was fixed as mentioned in the
previous paragraph.

The garbage collection module was explored to help figure this issue
out. Alrhough, the real method that lead to this solution was a print
statement displaying the number of skfuzzy sim runs had elapsed. It
remained 1 which lead to the fact that the control space was created
every single time. This was costly because skfuzzy uses about 300-400MB
of memory to create the control space.