Create ga for fis (#34)
* Modularized membership function value using numpy array

* Added membership function values as an optional argument

* Created test_ga script

* Using EasyGA module for genetic algorithm

* Added membership function passthrough as required variable to FIS

* Final attempt with PyGAD

* Scrappy code testing EasyGA

* Updated spreadsheet output naming

* Updated exception handling and more

Updated the use of thee error global variable that is set to true if
there is an exception.

Moved the population creation function outside of the My_GA() class so
the function can be used properly outside of the GA.

* Added custom survivor function

* Test parameters and saving to csv

* Commented ga graph and recnumber

* Gen 1000, Added timers and removed print statements

* Created anaconda environment file

* Cetch the error line number in exception

* Fixed memory leak issue

The skfuzzy control simulation was created four times per SUMO time
step. Each time the vehicles were calculating the acceleration. The
skfuzzy control space creation was moved and occurs once per chromosome.

The skfuzzy control holds cache for up to 100 simulations. This can be
expanded now that the real memory issue was fixed as mentioned in the
previous paragraph.

The garbage collection module was explored to help figure this issue
out. Alrhough, the real method that lead to this solution was a print
statement displaying the number of skfuzzy sim runs had elapsed. It
remained 1 which lead to the fact that the control space was created
every single time. This was costly because skfuzzy uses about 300-400MB
of memory to create the control space.

* Additional fixes following skfuzzy memory leak fix

Adjusted the ga sumo script to properly use the updated fuzzy
controller. This now creates the skfuzzy control space once per
chromosome instead of four times a sumo timestep.

* Updated the non ga sumo script following memory leak fix

* Plot lowest fitness values

* Reorganized use of class

* Sumo scrip runs with different chromosome

* Added plotting functionality back in

* Updated title back to Krauss from CACC

* File changes used to inform changes to the main test_eas_ga.py script

* Copied over working fis/gfs files to src