Update gap error rate (#36)
* Updated gap error membership functions

* Updated training membership functions and removed vehicle 5

The hand picked membership functions for the training scripts were
updated. The fifth vehicle was removed simplifying the training and the
simulation.

* Made the cacc model more cooperative and strategic

* Several modifications and additions

Modified the slowdown speed when a slowdown is supposed to occur. The
vehicles did not reach the desired slow speed previously within the time
frmae.

Added the plotting of each vehicle time game.

Added units to plots

* Updated gap error rate calculation

* updated first fis mem for lat ga